Fix handling of combining characters and control characters
The ANSEL specification describes combining characters as:

> Nonspacing graphic character (combining character)-A graphic character whose
> use is not followed by the forward movement of the output device. For the
> purpose of this standard, the term includes character modifiers.

Given that interpretation, when a combining character is written to the output
stream, and isn't followed by a spacing character, it should implicitly be
treated as if a space character (`U+0020`) was written to the output device.
When this is translated to unicode, it will result in the combining characters
following the space character, which when rendered will appear as if the
combining characters alone were printed.

Fixes #9